Refactor registration using abstract class.
